# TangWebApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TangWebApi",
    "description": "基于.NET 8的Web API项目，集成SqlSugar、Swagger和Knife4j",
    "contact": {
      "name": "TangWebApi",
      "email": "admin@tangwebapi.com"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Logging/debug": {
      "post": {
        "tags": [
          "日志测试"
        ],
        "summary": "测试调试日志",
        "operationId": "TestDebugLog (Logging)",
        "requestBody": {
          "description": "日志消息",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Logging/info": {
      "post": {
        "tags": [
          "日志测试"
        ],
        "summary": "测试信息日志",
        "operationId": "TestInfoLog (Logging)",
        "requestBody": {
          "description": "日志消息",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Logging/warning": {
      "post": {
        "tags": [
          "日志测试"
        ],
        "summary": "测试警告日志",
        "operationId": "TestWarningLog (Logging)",
        "requestBody": {
          "description": "日志消息",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Logging/error": {
      "post": {
        "tags": [
          "日志测试"
        ],
        "summary": "测试错误日志",
        "operationId": "TestErrorLog (Logging)",
        "requestBody": {
          "description": "日志消息",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Logging/critical": {
      "post": {
        "tags": [
          "日志测试"
        ],
        "summary": "测试严重错误日志",
        "operationId": "TestCriticalLog (Logging)",
        "requestBody": {
          "description": "日志消息",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Logging/structured": {
      "post": {
        "tags": [
          "日志测试"
        ],
        "summary": "测试结构化日志",
        "operationId": "TestStructuredLog (Logging)",
        "requestBody": {
          "description": "结构化日志请求",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructuredLogRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StructuredLogRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StructuredLogRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Logging/business": {
      "post": {
        "tags": [
          "日志测试"
        ],
        "summary": "测试业务操作日志",
        "operationId": "TestBusinessLog (Logging)",
        "requestBody": {
          "description": "业务操作请求",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessLogRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessLogRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessLogRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Logging/performance/{delayMs}": {
      "post": {
        "tags": [
          "日志测试"
        ],
        "summary": "测试性能日志（模拟慢操作）",
        "operationId": "TestPerformanceLog (Logging)",
        "parameters": [
          {
            "name": "delayMs",
            "in": "path",
            "description": "延迟毫秒数",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Logging/config": {
      "get": {
        "tags": [
          "日志测试"
        ],
        "summary": "获取日志配置信息",
        "operationId": "GetLoggingConfig (Logging)",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BusinessLogRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "operation": {
            "type": "string",
            "nullable": true
          },
          "resource": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "业务操作日志请求模型"
      },
      "LogLevel": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "description": "日志级别枚举",
        "format": "int32"
      },
...
      "StructuredLogRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "action": {
            "type": "string",
            "nullable": true
          },
          "resource": {
            "type": "string",
            "nullable": true
          },
          "logLevel": {
            "$ref": "#/components/schemas/LogLevel"
          }
        },
        "additionalProperties": false,
        "description": "结构化日志请求模型"
      },
...
    },
    "securitySchemes": {
      "Authorization": {
        "type": "apiKey",
        "description": "JWT授权token前面需要加上字段Bearer与一个空格,如Bearer token",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}